djangoapp/requirements.txt
Django==3.1.*
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.0
markdown
django-filter
django-guardian
Pygments
python-dotenv
gunicorn
autopep8
django-sass
Pillow
django-cors-headers
pyyaml
uritemplate
psycopg2-binary
drf-yasg==1.20.0